+++ packages/ethers-ethereum/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.2](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.7.1...v0.7.2) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/ethers-ethereum
+
+
+
+
+
+# [0.7.0](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.6.3...v0.7.0) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/ethers-ethereum
+
+
+
+
+
 # [0.6.0](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.5.4...v0.6.0) (2021-10-07)
 
 **Note:** Version bump only for package @rarible/ethers-ethereum