+++ packages/sdk-bundle/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.4](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.7.3...v0.7.4) (2021-10-14)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+## [0.7.3](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.7.2...v0.7.3) (2021-10-14)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+## [0.7.2](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.7.1...v0.7.2) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+## [0.7.1](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.7.0...v0.7.1) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+# [0.7.0](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.6.3...v0.7.0) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+## [0.6.3](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.6.2...v0.6.3) (2021-10-13)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
+## [0.6.2](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.6.1...v0.6.2) (2021-10-11)
+
+**Note:** Version bump only for package @rarible/sdk-bundle
+
+
+
+
+
 ## [0.6.1](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.6.0...v0.6.1) (2021-10-11)
 
 **Note:** Version bump only for package @rarible/sdk-bundle